separate StepIndicator
