+++ app/components/indicators/index.js
@@ -13,7 +13,7 @@ export { default as LinearProgressSmall } from "./LinearProgress/LinearProgressS
 export { default as LoaderBarBottom } from "./LoaderBarBottom";
 export { default as NoStats } from "./NoStats";
 export { default as NoTransactions } from "./NoTransactions";
-export { default as StepIndicator } from "./StepIndicator";
+export { default as StepIndicator } from "./StepIndicator/StepIndicator";
 export { default as VotingProgress } from "./VotingProgress";
 export { default as NoProposals } from "./NoProposals";
 export { default as PoliteiaLoading } from "./PoliteiaLoading";
